+++ src/types/sdk/ClientConfiguration.ts
@@ -2,4 +2,5 @@ export interface ClientConfiguration {
   bearerToken: string;
   baseUrl?: string;
   stelaBaseUrl?: string;
+  retryOn?: number[];
 }